Move fns to different file
